Rename instance.rb to base.rb
