+++ lib/pansophy_authenticator/configuration.rb
@@ -6,5 +6,5 @@ end
 require 'pansophy_authenticator/configuration/configurator'
 require 'pansophy_authenticator/configuration/from_file'
 require 'pansophy_authenticator/configuration/from_env'
-require 'pansophy_authenticator/configuration/instance'
+require 'pansophy_authenticator/configuration/base'
 require 'pansophy_authenticator/configuration/verification'
