+++ lib/pansophy_authenticator/configuration/base.rb
